Improved code readibility to my modifications.
